fix: updates per design discussion
